# contracts/http/belcanto-v1.openapi.yaml
openapi: 3.1.0
...
info:
  title: Belcanto API
  version: 0.4.0
  description: >-
    HTTP contract for closed B.0 onboarding, L.1 internal Lesson scheduling
    and P.1 session security (device inventory, targeted revocation, the
    account security feed and one-time password recovery).
    There is intentionally no public sign-up operation: the school creates a
    Student and the invited person only activates the prepared Account.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/me/profile:
    get:
      summary: Read the caller's profile
      operationId: myProfile
      tags: [Security]
      description: Display identity for the account area (ACC-01/02).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile identity with tenant and role context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileView"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/Internal"
    put:
      summary: Update the caller's profile
      operationId: updateMyProfile
      tags: [Security]
      description: >-
        Changes the display name only; contact values change through the
        verified contact-change flow.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: The updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileView"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    InvalidInput:
      description: Request shape or value is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InvalidInputError"
...
    Unauthenticated:
      description: Authentication or fresh credential verification failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UnauthenticatedError"
...
    Internal:
      description: An unexpected server failure occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InternalError"
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - INVALID_INPUT
        - UNAUTHENTICATED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INVALID_STATE
        - ACTIVATION_INVALID
        - RATE_LIMITED
        - UNAVAILABLE
        - INTERNAL
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              minLength: 1
            requestId:
              type: string
              minLength: 1
    InvalidInputError:
      allOf:
        - $ref: "#/components/schemas/ErrorEnvelope"
        - properties:
            error:
              properties:
                code:
                  const: INVALID_INPUT
...
    UnauthenticatedError:
      allOf:
        - $ref: "#/components/schemas/ErrorEnvelope"
        - properties:
            error:
              properties:
                code:
                  const: UNAUTHENTICATED
...
    InternalError:
      allOf:
        - $ref: "#/components/schemas/ErrorEnvelope"
        - properties:
            error:
              properties:
                code:
                  const: INTERNAL
...
    ProfileView:
      type: object
      additionalProperties: false
      required: [accountId, fullName, tenantName, roles, phone]
      properties:
        accountId:
          type: string
          minLength: 1
          maxLength: 128
        fullName:
          type: string
          minLength: 1
          maxLength: 200
        tenantName:
          type: string
          minLength: 1
          maxLength: 200
        roles:
          type: array
          items:
            type: string
            enum: [Owner, Administrator, Teacher, Student]
        phone:
          type: string
          maxLength: 20
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      required: [fullName]
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 200
...